requirements.txt
ariadne==0.13.0
click==8.0.1
Flask==2.0.1
Flask-PyMongo==2.3.0
graphql-core==3.1.5
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pymongo==3.11.4
python-dotenv==0.18.0
starlette==0.14.2
typing-extensions==3.10.0.0
Werkzeug==2.0.1
